New emit fieldValueChanged returning array(id, value)
Needed to know the value of an input field in javascript when it was changed. This Emit returns an array with the field_id and the value of what the field currently is. It also waits 700ms before issuing an emit to ensure the user is done typing.

It's ok if you don't want to use this addition. It works great for my needs. Maybe there already is something that does this, but I couldn't find it. I'm new to Vue.